requirements.txt
openai>=1.0.0
gitpython>=3.1.0
click>=8.0.0
pyyaml>=6.0.0
pytest>=7.0.0  # for testing
black>=22.0.0  # for code formatting
flake8>=4.0.0  # for linting
build>=0.7.0  # for building package
twine>=4.0.0  # for publishing to PyPI
